+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+### 0.4.0
+
+* Add the ability to watch the project using the `flutter pub run build_runner watch` command.
+* Improve class generation
+* Improve building with `build_runner`
+
 ### 0.2.0
 
 Added a new command that allows to run `build_runner`'s `build` command easily, just press `Ctrl+Shift+B`.
@@ -7,7 +13,6 @@ Added a new command that allows to run `build_runner`'s `build` command easily, 
 ### 0.1.0
 
 Added the ability to generate the file from the Explorer View. Right click in a folder on the right and select : `Freezed : Generate a new Freezed class`.
-
 
 ### 0.0.1